pyproject.toml
[tool.poetry]
name = "shi"
version = "0.1.0"
description = ""
authors = ["Qiyas <qiyasturk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-modeltranslation = "^0.19.11"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
